#661 PerpetualStreamMatValue.matValue should support PerpetualStreams (#692)
*  #661 PerpetualStreamMatValue.matValue should support PerpetualStreams

that materialize to `Sink[T, NotUsed]` or to any `Product` whose first element
is a `Sink[T, NotUsed]`.

*  #613 Update comment with @anilgursel suggestion.

*  #613 Handled akka.japi.Pair and java.util.List cases...

Moved the PerpetualStreamMatValue tests to a separate file.  I used
FunSpec over FlatSpec because I'm super familiar with reducing code and
producing a nice report using FunSpec.  It wasn't obvious how to port
this style to FlatSpec..

*  #613 Cover success cases for akka.japi.Pair and java.util.List

*  #613 Cleanup based on Akara's suggestions, plus...

a significant rework to the test actor's `receive l: Long` such that it
actually does what I think it does.  Note that there is still some
dependency on hoping that enough time has passed for things to be in the
correct state:(

* Remove unused `Inside` trait.

* Incorporate Akara's suggestions.  My only addition was to terminate the..

actor system in an afterAll override.